Rename internals to _internal
